Move ChainingTransformer to only package that it's used in
